[SPARK-28891][BUILD][2.3] backport do-release-docker.sh to branch-2.3
### What changes were proposed in this pull request?
This PR re-enables `do-release-docker.sh` for branch-2.3.

According to the release manager of Spark 2.3.3 maropu, `do-release-docker.sh` in the master branch. After applying #23098, the script does not work for branch-2.3.

### Why are the changes needed?
This PR simplifies the release process in branch-2.3 simple.

While Spark 2.3.x will not be released further, as dongjoon-hyun [suggested](https://github.com/apache/spark/pull/23098#issuecomment-524682234), it would be good to put this change for

1. to reproduce this release by others
2. to make the future urgent release simple

### Does this PR introduce any user-facing change?
No

### How was this patch tested?
No test is added.
This PR is used to create Spark 2.3.4-rc1

Closes #25607 from kiszk/SPARK-28891.

Authored-by: Kazuaki Ishizaki <ishizaki@jp.ibm.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>